requirements.txt
fastapi
uvicorn
langchain
openai
faiss-cpu
tiktoken
python-dotenv